refactor(promql): move guide to pkg/promql, add smoke test, fix bucket suffix
Three follow-ups from audit review:

1. Package home — moved the PromQL guide constant out of pkg/dashboard/query.go
   into a new pkg/promql package as promql.Instructions. The only consumer
   (internal/handler/tools/dashboards.go) switched to the new import. Dashboard-
   specific constants (ClickHouse SQL examples, Querybuilder) stay where they are.
   URI (signoz://promql/instructions) now aligns with the package name.

2. Registration smoke test — TestIntegration_PromqlInstructionsResourceRegistered
   in internal/mcp-server/integration_test.go reads signoz://promql/instructions
   via the in-process MCP client and asserts the body carries three load-bearing
   substrings ("Prometheus 3.x UTF-8 quoted selector", "payment_latency_ms.bucket",
   "group_right"). Catches future accidental rename or constant drift.

3. Bucket-suffix fix — review flagged http.server.duration_bucket as
   inconsistent with the guide's own histogram-suffix rule (dotted base must
   pair with .bucket dot-suffix, not _bucket). Corrected to
   http.server.duration.bucket.
